requirements.txt
Flask==1.1.2
GitPython==3.1.12
uWSGI==2.0.19.1
python-dotenv==0.19.0
